requirements.txt
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
html5lib==1.1
httplib2==0.22.0
numpy==1.26.4
pandas==2.2.1
plotly==5.20.0
python-dateutil==2.9.0.post0
scipy==1.12.0
sympy==1.12
yfinance==0.2.37

